+++ src/Paket.Core/Dependencies/NuGetV2.fs
@@ -73,7 +73,11 @@ let private followODataLink auth url =
 
             match atLeastOneFailed with
             | Some i ->
-                traceWarnfn "At least one 'next' link (index %d) returned a empty result (noticed on '%O'): ['%s']" i url (System.String.Join("' ; '", linksToFollow))
+                let mutable uri = null // warn once per specific API endpoint, but try to cut the query
+                let baseUrl = if Uri.TryCreate(url, UriKind.Absolute, &uri) then uri.AbsolutePath else url
+                traceWarnIfNotBefore baseUrl
+                    "At least one 'next' link (index %d) returned a empty result (noticed on '%O'): ['%s']" 
+                    i url (System.String.Join("' ; '", linksToFollow))
             | None -> ()
             return
                 true,
@@ -98,7 +102,7 @@ let tryGetAllVersionsFromNugetODataWithFilter (auth, nugetURL, package:PackageNa
                 match tryGetAllVersionsFromNugetODataWithFilterWarnings.TryGetValue nugetURL with
                 | true, true -> ()
                 | _, _ ->
-                    eprintfn "Possible Performance degradation, could not retrieve '%s', ignoring further warnings for this source" url
+                    traceWarnfn "Possible Performance degradation, could not retrieve '%s', ignoring further warnings for this source" url
                     tryGetAllVersionsFromNugetODataWithFilterWarnings.TryAdd(nugetURL, true) |> ignore
                 if verbose then
                     printfn "Error while retrieving data from '%s': %O" url exn
@@ -213,7 +217,8 @@ let private handleODataEntry nugetURL packageName version entry =
     let dependencies, warnings =
         addFrameworkRestrictionsToDependencies cleanedPackages frameworks
     for warning in warnings do
-        Logging.traceWarnfn "%s" (warning.Format officialName version)
+        let message = warning.Format officialName version
+        Logging.traceWarnIfNotBefore message "%s" message
 
     { PackageName = officialName
       DownloadUrl = downloadLink